userService/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
redis
passlib
bcrypt
python-jose
grpcio
grpcio-tools
